+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - MakeAvail: trigger update satellite if only FLAGS changed
 - Resource delete: Don't check user set verify-algo if last resource deleted.
 - If mkfs failed, resources have not been demoted(secondary) and couldn't easily be deleted afterward
+- Fixed incorrect least common multiple call when calculating AllocationGranularity
 
 ## [1.27.0] - 2024-04-02